fix(otel): build shared runtime on the exporter's dedicated thread
Construct the shared BasicRuntime inside build_on_dedicated_thread (the
exporter's own std thread) instead of in DatadogSpanProcessor::new on the
caller's thread. The runtime therefore never exists on the caller's thread, so
no exporter-init failure path can drop a tokio runtime inline there — dropping a
runtime runs a blocking shutdown that panics when new() is invoked from within a
tokio runtime, and that panic (via the public builder's catch_unwind) discarded
any user-supplied provider.

This covers both failure modes uniformly: a runtime build/exporter build failure
drops the runtime on the dedicated thread, and a thread-spawn failure never
creates a runtime at all. On success the exporter remains the runtime's sole
owner and drops it off-thread via Drop for DatadogExporter.

Remote Config and telemetry are now registered after the exporter is built,
through span_exporter.shared_runtime() (widened to pub(crate)).
